# docs/Collections/swagger.yaml
openapi: 3.0.2
info:
  title: "Collections"
  version: 1.0.0
  description: This is a simple Collections microservice API by team 3
paths:
  /collections:
    post:
      tags: 
        - "Create"
      summary: "Create new collection"
      description: "Create new collection and get collection id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                token:
                  type: "string"
                name:
                  type: "string"
                description:
                  type: "string"
                firstPostId:
                  type: "integer"
                  format: "int64"
                criterion:
                  type: "array"
                  items:
                    type: "string"
                    format: "string"
            example:
              token: "f57ebe597a3741b688269209fa29b053"
              name: "Чайники"
              description: "Сравнение моих чайников"
              first_post_id: 322
              criterion: ["Качество", "Цена", "Внешний вид"]
      responses:
        201:
          description: "Created"
        401:
          description: "Not enough rights to create the collection"
        404:
          description: "Error"
          
    put:
      tags:
        - "Edit"
      summary: "Edit collection data"
      description: "Edit name, description, criterion list"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                token:
                  type: "string"
                collectionID:
                  type: "integer"
                  format: "int64"
                name:
                  type: "string"
                description:
                  type: "string"
                criterion:
                  type: "array"
                  items:
                    type: "string"
                    format: "string"
            example:
              token: "f57ebe597a3741b688269209fa29b053"
              collection_id: 318
              name: "Чайники"
              description: "Сравнение моих чайников"
              criterion: ["Качество", "Цена", "Внешний вид"]
      responses:
        200:
          description: "OK"
        401:
          description: "Not enough rights to create the collection"
        404:
          description: "Error"
          
      
  
  /collections/posts:
      put:
        tags: 
          - "Edit"
        summary: "Add post to collection"
        description: ""
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  token:
                    type: "string"
                  collection_id:
                    type: "integer"
                    format: "int64"
                  postID:
                    type: "integer"
                    format: "int64"
              example:
                token: "f57ebe597a3741b688269209fa29b053"
                collection_id: "1445"
                postID: 333
        responses:
          200:
            description: "OK"
          401:
            description: "Not enough rights to add a post to the collection"
          404:
            description: "Error"
            
  /collections/posts/{postID}/token/{token}:
    delete:
      tags:
        - "Delete"
      summary: "Remove one post from all collections"
      description: ""
      parameters:
      - name: "token"
        in: "path"
        description: "User token"
        required: true
        schema:
          type: "string"
      - name: "postID"
        in: "path"
        description: "ID post"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "OK"
        401:
          description: "Not enough rights"
        404:
          description: "Error"
          
  /collections/posts/token/{token}:
    delete:
      tags:
        - "Delete"
      summary: "Remove posts from all collections.
       Posts to be deleted are sent as a list"
      description: ""
      parameters:
      - name: "token"
        in: "path"
        description: "User token"
        required: true
        schema:
          type: "string"
      - name: "postsList"
        in: "query"
        description: "List of users IDs"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
        example: 
          [11, 45, 99, 85]
      responses:
        200:
          description: "OK"
        401:
          description: "Not enough rights"
        404:
          description: "Error"
      
  
  /collections/{collectionID}/posts:
      get:
        tags: 
          - "Get"
        summary: "Get a list of post IDs in a collection"
        description: ""
        parameters:
        - name: "collectionID"
          in: "path"
          description: "Collection Id"
          required: true
          schema:
            type: "integer"
            format: "int64"
        responses:
          200:
            description: "OK"
            content: 
              pplication/json:
                schema:
                  type: "object"
                  properties:
                    posts:  
                      type: "array"
                      items:
                        format: "integer"
                example:
                  posts: [322, 333, 5, 4692]
          404:
            description: "Not found"
                  
  /collections/{collectionID}/token/{token}:
      get:
        tags: 
        - "Get"
        summary: "Get collection data"
        description: ""
        parameters:
        - name: "token"
          in: "path"
          description: "User token"
          required: true
          schema:
            type: "string"
        - name: "collectionID"
          in: "path"
          description: "ID of collection"
          required: true
          schema:
            type: "string"
        responses:
          200:
            description: "OK"
            content:
              application/json:
                schema:
                  type: "object"
                  properties:
                    name:
                      type: "string"
                    owner:
                      type: "integer"
                    discription:
                      type: "string"
                    posts: 
                      type: "array"
                      items: 
                        type: "integer"
                        format: "int64"
                    criterion: 
                      type: "array"
                      items: 
                        type: "string"
                example:
                  name: "Чайники"
                  owner: 18
                  description: "Сравнение моих чайников"
                  posts: [322, 777]
                  criterion: ["Качество", "Цена", "Внешний вид"]
                  
          401:
            description: "Not enough rights"
            content: {}
          
          404:
            description: "Error"
            
      delete:
        tags:
          - "Delete"
        summary: "Remove collection"
        description: ""
        parameters:
        - name: "token"
          in: "path"
          description: "User token"
          required: true
          schema:
            type: "string"
        - name: "collectionID"
          in: "path"
          description: "ID collection"
          required: true
          schema:
            type: "integer"
            format: "int64"
        responses:
          200:
            description: "OK"
          401:
            description: "Not enough rights"  
          
          404:
            description: "Non-existing token or non-existing or ID collection"
            
  /collections/users/{userID}/token/{token}:
    get:
      tags: 
        - "Get"
      summary: "Get all user collections"
      description: "Getting a list of collections owned by a user"
      parameters:
      - name: "token"
        in: "path"
        description: "User token"
        required: true
        schema:
          type: "string"
      - name: "userID"
        in: "path"
        description: "User ID"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                  type: "object"
                  properties:
                    posts:  
                      type: "array"
                      items:
                        format: "integer"
              example:
                collections: [322, 333, 5, 4692]
        401:
            description: "Not enough rights"
            content: {}
        
        404:
          description: "Error"
  
  /collections/{collectionID}/posts/{postID}/token/{token}:
    delete:
      tags:
        - "Delete"
      summary: "Remove post from collection"
      description: ""
      parameters:
      - name: "token"
        in: "path"
        description: "User token"
        required: true
        schema:
          type: "string"
      - name: "collectionID"
        in: "path"
        description: "ID collection"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "postID"
        in: "path"
        description: "ID post"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "OK"
        401:
          description: "Not enough rights"
        404:
          description: "Non-existing token or non-existing ID post or ID collection"